+++ doc/changelog.md
@@ -4,7 +4,7 @@
   - Better log message when waiting for URL (#640)
   - Extended authentication for AWS ECR (#663)
   - Add two new goals: "volume-create" and "volume-remove" for volume handling independent of images. 
-
+  - Support for loading from an tar archive (option `<build><dockerArchive>`) (#645)
 
 * **0.18.1** (2016-11-17)
   - Renamed `basedir` and `exportBasedir` in an `<assembly>` configuration to `targetDir` and `exportTargetDir` since this better reflects the purpose, i.e. the target in the Docker image to which the assembly is copied. The old name is still recognized but deprecated.